requirements.txt
streamlit==1.29.0
pandas==2.0.3
Pillow==9.4.0